+++ src/dashboard/dto/dashboard-state-response.dto.ts
