+++ src/embedding/dto/embedding-preview-request.dto.ts
@@ -2,13 +2,13 @@ import type {
   EmbeddingPreviewLanguage,
   EmbeddingPreviewSize,
   EmbeddingPreviewTheme,
-} from './embedding-preview.response';
+} from './embedding-preview-response.dto';
 
 export type {
   EmbeddingPreviewLanguage,
   EmbeddingPreviewSize,
   EmbeddingPreviewTheme,
-} from './embedding-preview.response';
+} from './embedding-preview-response.dto';
 
 export interface EmbeddingPreviewQueryDto {
   userId: string;
