+++ src/embedding/dto/embedding-preview-response.dto.ts
