+++ src/inventory/dto/inventory-request.dto.ts
