+++ src/inventory/dto/inventory-response.dto.ts
