+++ src/level-up/dto/level-up-request.dto.ts
@@ -1,4 +1,4 @@
-import type { LevelUpStat } from './level-up.response';
+import type { LevelUpStat } from './level-up-response.dto';
 
 export interface LevelUpApplyRequest {
   stat: LevelUpStat;
