+++ src/logs/logs.service.ts
@@ -5,7 +5,7 @@ import { encodeLogsCursor } from './logs-cursor.util';
 import type { ValidatedLogsQuery } from './logs-query.validator';
 import type { DungeonLogDelta } from '../common/logs/dungeon-log-delta';
 import type { DungeonLogDetails } from '../common/logs/dungeon-log-extra';
-import type { DungeonLogsPayload } from './dto/logs.response';
+import type { DungeonLogsPayload } from './dto/logs-response.dto';
 import { isLogAction, type LogAction, type LogStatus } from './logs.types';
 
 @Injectable()
