+++ src/ranking/dto/ranking-query.dto.ts
