+++ src/ranking/ranking.controller.ts
@@ -5,7 +5,7 @@ import { successResponseWithGeneratedAt } from '../common/http/api-response';
 import { applyNoCacheHeaders } from '../common/http/response-helpers';
 import { AuthenticatedThrottlerGuard } from '../common/guards/authenticated-throttler.guard';
 import { validateRankingQuery } from './ranking-query.validator';
-import type { RankingPayload } from './dto/ranking.response';
+import type { RankingPayload } from './dto/ranking-response.dto';
 import { RankingService } from './ranking.service';
 
 @Controller('api')
