+++ src/settings/settings.service.ts
@@ -7,7 +7,7 @@ import {
 import type { ActiveSessionResult } from '../auth/auth-session.service';
 import { RuntimeValidationError } from '../common/validation/runtime-validation';
 import { PrismaService } from '../prisma/prisma.service';
-import type { SettingsProfileResponse } from './dto/settings-profile.response';
+import type { SettingsProfileResponse } from './dto/settings-profile-response.dto';
 import { assertSettingsProfileResponse } from './settings-profile.validator';
 
 @Injectable()
